refactor tests to smaller batches
