+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split4/VerifyLightGBMClassifierStream.scala
@@ -1,86 +1,26 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lightgbm.split1
+package com.microsoft.azure.synapse.ml.lightgbm.split4
 
 import com.microsoft.azure.synapse.ml.lightgbm._
+import com.microsoft.azure.synapse.ml.lightgbm.dataset.LightGBMDataset
+import com.microsoft.azure.synapse.ml.lightgbm.params.FObjTrait
+import com.microsoft.azure.synapse.ml.lightgbm.split1.LightGBMClassifierTestData
 import org.apache.spark.TaskContext
 import org.apache.spark.ml.feature.LabeledPoint
 import org.apache.spark.ml.linalg.{DenseVector, Vectors}
-import org.apache.spark.ml.tuning.{ParamGridBuilder, TrainValidationSplit}
 import org.apache.spark.sql.catalyst.encoders.RowEncoder
 import org.apache.spark.sql.functions._
 
+import scala.math.exp
+
 // scalastyle:off magic.number
+
 /** Tests to validate the functionality of LightGBM module in streaming mode. */
 class VerifyLightGBMClassifierStream extends LightGBMClassifierTestData {
-  /* TODO Figure out why abalone has such poor score
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.MulticlassObjective, abaloneFile, executionMode)) {
-    verifyLearnerOnMulticlassCsvFile(abaloneFile, "Rings", 2)
-  } */
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.MulticlassObjective, breastTissueFile, executionMode)) {
-    verifyLearnerOnMulticlassCsvFile(breastTissueFile, "Class", .07)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.MulticlassObjective, carEvaluationFile, executionMode)) {
-    verifyLearnerOnMulticlassCsvFile(carEvaluationFile, "Col7", 2)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, pimaIndianFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(pimaIndianFile, "Diabetes mellitus", 1)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, banknoteFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(banknoteFile, "class", 1)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, taskFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(taskFile, "TaskFailed10", 1)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, breastCancerFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(breastCancerFile, "Label", 1)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, randomForestFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(randomForestFile, "#Malignant", 1)
-  }
-  test(verifyLearnerTitleTemplate.format(LightGBMConstants.BinaryObjective, transfusionFile, executionMode)) {
-    verifyLearnerOnBinaryCsvFile(transfusionFile, "Donated", 1)
-  }
-
-  test("Verify LightGBMClassifier save booster to " + pimaIndianFile + executionModeSuffix) {
-    verifySaveBooster(
-      fileName = pimaIndianFile,
-      labelColumnName = "Diabetes mellitus",
-      outputFileName = "model.txt",
-      colsToVerify = Array("Diabetes pedigree function", "Age (years)"))
-  }
-
-  test("Compare benchmark results file to generated file" + executionModeSuffix) {
-    verifyBenchmarks()
-  }
-
-  test("Verify LightGBM Classifier can be run with TrainValidationSplit" + executionModeSuffix) {
-    val model = baseModel.setUseBarrierExecutionMode(true)
-
-    val paramGrid = new ParamGridBuilder()
-      .addGrid(model.numLeaves, Array(5, 10))
-      .addGrid(model.numIterations, Array(10, 20))
-      .addGrid(model.lambdaL1, Array(0.1, 0.5))
-      .addGrid(model.lambdaL2, Array(0.1, 0.5))
-      .build()
-
-    val fitModel = new TrainValidationSplit()
-      .setEstimator(model)
-      .setEvaluator(binaryEvaluator)
-      .setEstimatorParamMaps(paramGrid)
-      .setTrainRatio(0.8)
-      .setParallelism(2)
-      .fit(pimaDF)
-
-    fitModel.transform(pimaDF)
-    assert(fitModel != null)
-
-    // Validate lambda parameters set on model
-    val modelStr = fitModel.bestModel.asInstanceOf[LightGBMClassificationModel].getModel.modelStr.get
-    assert(modelStr.contains("[lambda_l1: 0.1]") || modelStr.contains("[lambda_l1: 0.5]"))
-    assert(modelStr.contains("[lambda_l2: 0.1]") || modelStr.contains("[lambda_l2: 0.5]"))
-  }
+  override def ignoreSerializationFuzzing: Boolean = true
+  override def ignoreExperimentFuzzing: Boolean = true
 
   test("Verify LightGBM Classifier with batch training" + executionModeSuffix) {
     val batches = Array(0, 2, 10)
@@ -100,7 +40,7 @@ class VerifyLightGBMClassifierStream extends LightGBMClassifierTestData {
     })
   }
 
-  /*test("Verify LightGBM Multiclass Classifier with vector initial score" + executionModeSuffix) {
+  test("Verify LightGBM Multiclass Classifier with vector initial score" + executionModeSuffix) {
     val multiClassModel =
       baseModel.setObjective(LightGBMConstants.MulticlassObjective).setSeed(4).setDeterministic(true)
     val scoredDF1 = multiClassModel.fit(breastTissueDF).transform(breastTissueDF)
@@ -150,7 +90,7 @@ class VerifyLightGBMClassifierStream extends LightGBMClassifierTestData {
     assertBinaryImprovement(scoredDF1, scoredDF2)
   }
 
-  test("Verify LightGBM Classifier will give reproducible results when setting seed") {
+  /*test("Verify LightGBM Classifier will give reproducible results when setting seed") {
     val scoredDF1 = baseModel.setSeed(1).setDeterministic(true).fit(pimaDF).transform(pimaDF)
     (1 to 10).foreach { _ =>
       val scoredDF2 = baseModel.setSeed(1).setDeterministic(true).fit(pimaDF).transform(pimaDF)
